requirements.txt
fastapi
uvicorn
python-dotenv
google-genai
sqlalchemy
pg8000